pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-document-intelligence"
version = "0.1.0"
description = "Production-style Retrieval Augmented Generation pipeline for document intelligence"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.6",
    "pydantic-settings>=2.1",
    "python-dotenv>=1.0",
    "pypdf>=4.2",
    "sentence-transformers>=2.2",
    "faiss-cpu>=1.7",
    "transformers>=4.38",
    "torch>=2.0",
    "pandas>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]

[tool.setuptools.packages.find]
where = ["."]
include = ["rag*"]
